dev.requirements.txt
requests>=2.31.0
urllib3>=1.26.7
httpx>=0.23.3
pydantic>=2.5.3
python-dateutil>=2.8.2
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pre-commit>=3.6.0
